// types/data-grid.ts
export interface CellSelectOption {
  label: string
  value: string
  icon?: React.FC<React.SVGProps<SVGSVGElement>>
  count?: number
}

export type CellOpts =
  | {
      variant: "short-text"
    }
  | {
      variant: "long-text"
    }
  | {
      variant: "number"
      min?: number
      max?: number
      step?: number
    }
  | {
      variant: "select"
      options: CellSelectOption[]
    }
  | {
      variant: "multi-select"
      options: CellSelectOption[]
    }
  | {
      variant: "checkbox"
    }
  | {
      variant: "date"
    }
  | {
      variant: "url"
    }
  | {
      variant: "file"
      maxFileSize?: number
      maxFiles?: number
      accept?: string
      multiple?: boolean
    }